Trader_Companion-main/requirements.txt
## Unified, Python 3.13 compatible dependency set.
## Pin exact versions for reproducibility.
## If updating, adjust here and rerun setup script.
Django==5.1.3
Flask==3.1.0
numpy==2.2.1
pandas==2.2.3
requests==2.32.3
yfinance==0.2.54
django-cors-headers==4.4.0
psutil==6.0.0
djangorestframework==3.15.2
Pillow==11.0.0
flask-cors==6.0.0
pygame==2.6.1
pydub==0.25.1